test: ZENKO-2866 fix DummyObjectStream
- fix the DummyObjectStream test helper _read implementation

- separate tests of the helper itself into a separate file to ease
  reuse outside the mocha test framework
